feat/PK-17: inventory 관련 비즈니스 로직 hook으로 분리
